requirements.txt
asgiref==3.2.10
certifi==2021.5.30
chardet==4.0.0
coreapi==2.3.3
coreschema==0.0.4
Django==3.1
django-filter==2.4.0
djangorestframework==3.12.4
idna==2.10
importlib-metadata==4.5.0
itypes==1.2.0
Jinja2==3.0.1
Markdown==3.3.4
MarkupSafe==2.0.1
pkg-resources==0.0.0
Pygments==2.9.0
pytz==2021.1
requests==2.25.1
sqlparse==0.4.1
typing-extensions==3.10.0.0
uritemplate==3.0.1
urllib3==1.26.5
zipp==3.4.1
